Properly write the output in operator basic_pipe for async_pipe
Pretty sure this was backwards? It makes no sense to check
`source` and `sink`, you only declared them four lines ago and
nothing touched them since.
This way, we do something that does make sense: returning an
invalid handle if the handle stored in this object is itself an
invalid handle.