fix exit code when running under asio
When using a boost::asio::io_context, `io_context_ref` is expected to
handle SIGCHLD, call waitpid and store the exit code of the process. In
addition to storing the exit code, it also handles calling
`on_exit_handler` on any `async_handler` instances (e.g. when using
`on_exit`).

However, there was a check that only set up the SIGCHLD handling when
there were any `async_handler` instances present. So when there are none
(which is usually the case), the exit code is also not set, resulting in
an invalid result when calling `exit_code()`. To fix this, the check is
simply removed, so now a SIGCHLD handler is always set up when an
`io_context` is passed.

This check was introduced in commit 7017150 (added specialization for
posix && added conditional for io_service_ref).

This fixes https://github.com/boostorg/process/issues/39